refactor(ebpf-extractor): extract library with testable run() function
Move all extractor logic from main.rs into lib.rs, matching the pattern
established by the rpc, log, and p2p extractors. The new lib.rs exposes
a public Args struct (with a new() constructor for programmatic use) and
a public async run() function that can be called from integration tests.

main.rs is reduced to arg parsing and a direct call to
ebpf_extractor::run(). No behavioural changes - the extractor runs
identically to before.
